Move get_num_triton_sampler_splits into triton_utils
